refactor: move object manager into dedicated module
